fix: address takeover readiness feedback
