+++ src/daemon/human-control.ts
@@ -1,43 +1,42 @@
-import fs from 'node:fs';
-import path from 'node:path';
 import { AppError } from '@agent-device/kernel/errors';
 import { deviceIdentityAliases } from '../core/lease-scope.ts';
-import type {
-  HumanControlHold,
-  HumanControlHoldInput,
-  HumanControlHoldScope,
+import type { HumanControlHold, HumanControlHoldInput } from './human-control-contract.ts';
+import {
+  normalizeHumanControlHoldId,
+  normalizeHumanControlHoldScope,
+  normalizeHumanControlReason,
+  normalizeHumanControlTtlMs,
 } from './human-control-contract.ts';
+import { cloneHumanControlHold, HumanControlStore } from './human-control-store.ts';
 
 export const HUMAN_CONTROL_HTTP_PREFIX = '/admin/human-control/holds';
-
-const MIN_HOLD_TTL_MS = 1_000;
-const MAX_HOLD_TTL_MS = 24 * 60 * 60_000;
 
 export class HumanControlRegistry {
   private readonly holds = new Map<string, HumanControlHold>();
   private readonly activeMutations = new Map<string, number>();
   private readonly idleWaiters = new Map<string, Set<() => void>>();
-  private readonly statePath: string | undefined;
+  private readonly store: HumanControlStore;
   private readonly now: () => number;
 
   constructor(options: { statePath?: string; now?: () => number } = {}) {
-    this.statePath = options.statePath;
+    this.store = new HumanControlStore(options.statePath);
     this.now = options.now ?? (() => Date.now());
-    this.load();
+    for (const hold of this.store.load()) this.holds.set(hold.id, hold);
+    this.cleanupExpired();
   }
 
   list(): HumanControlHold[] {
     this.cleanupExpired();
-    return Array.from(this.holds.values(), (hold) => cloneHold(hold)).sort((left, right) =>
-      left.id.localeCompare(right.id),
+    return Array.from(this.holds.values(), (hold) => cloneHumanControlHold(hold)).sort(
+      (left, right) => left.id.localeCompare(right.id),
     );
   }
 
   async upsert(id: string, input: HumanControlHoldInput): Promise<HumanControlHold> {
-    const normalizedId = normalizeHoldId(id);
-    const scope = normalizeScope(input.scope);
-    const reason = normalizeReason(input.reason);
-    const ttlMs = normalizeTtlMs(input.ttlMs);
+    const normalizedId = normalizeHumanControlHoldId(id);
+    const scope = normalizeHumanControlHoldScope(input.scope);
+    const reason = normalizeHumanControlReason(input.reason);
+    const ttlMs = normalizeHumanControlTtlMs(input.ttlMs);
     const now = this.now();
     const existing = this.holds.get(normalizedId);
     const pendingHold: HumanControlHold = {
@@ -65,16 +64,16 @@ export class HumanControlRegistry {
     };
     this.holds.set(normalizedId, activeHold);
     this.persist();
-    return cloneHold(activeHold);
+    return cloneHumanControlHold(activeHold);
   }
 
   remove(id: string): HumanControlHold | undefined {
-    const normalizedId = normalizeHoldId(id);
+    const normalizedId = normalizeHumanControlHoldId(id);
     const hold = this.holds.get(normalizedId);
     if (!hold) return undefined;
     this.holds.delete(normalizedId);
     this.persist();
-    return cloneHold(hold);
+    return cloneHumanControlHold(hold);
   }
 
   isDeviceControlled(deviceKey: string | undefined): boolean {
@@ -88,7 +87,7 @@ export class HumanControlRegistry {
     if (keys.length === 0) return undefined;
     for (const hold of this.holds.values()) {
       const holdKeys = normalizeDeviceAliases([hold.scope.deviceKey]);
-      if (holdKeys.some((key) => keys.includes(key))) return cloneHold(hold);
+      if (holdKeys.some((key) => keys.includes(key))) return cloneHumanControlHold(hold);
     }
     return undefined;
   }
@@ -148,42 +147,8 @@ export class HumanControlRegistry {
     if (changed) this.persist();
   }
 
-  private load(): void {
-    if (!this.statePath || !fs.existsSync(this.statePath)) return;
-    let parsed: { version: 1; holds: HumanControlHold[] };
-    try {
-      parsed = JSON.parse(fs.readFileSync(this.statePath, 'utf8')) as typeof parsed;
-    } catch (error) {
-      throw new AppError(
-        'COMMAND_FAILED',
-        'Failed to read persisted human-control state.',
-        { path: this.statePath },
-        error,
-      );
-    }
-    if (parsed.version !== 1 || !Array.isArray(parsed.holds)) {
-      throw new AppError('COMMAND_FAILED', 'Persisted human-control state is invalid.', {
-        path: this.statePath,
-      });
-    }
-    for (const rawHold of parsed.holds) {
-      const hold = normalizeStoredHold(rawHold);
-      this.holds.set(hold.id, hold);
-    }
-    this.cleanupExpired();
-  }
-
   private persist(): void {
-    if (!this.statePath) return;
-    fs.mkdirSync(path.dirname(this.statePath), { recursive: true });
-    const temporaryPath = `${this.statePath}.${String(process.pid)}.tmp`;
-    const state = {
-      version: 1,
-      holds: Array.from(this.holds.values(), (hold) => cloneHold(hold)),
-    };
-    fs.writeFileSync(temporaryPath, JSON.stringify(state, null, 2), { mode: 0o600 });
-    fs.renameSync(temporaryPath, this.statePath);
-    fs.chmodSync(this.statePath, 0o600);
+    this.store.persist(this.holds.values());
   }
 }
 
@@ -213,59 +178,6 @@ export function releaseHumanControlHold(
   return registry.remove(holdId);
 }
 
-function normalizeStoredHold(raw: HumanControlHold): HumanControlHold {
-  if (!raw || typeof raw !== 'object') {
-    throw new AppError('COMMAND_FAILED', 'Persisted human-control hold is invalid.');
-  }
-  const createdAt = normalizeTimestamp(raw.createdAt, 'createdAt');
-  const updatedAt = normalizeTimestamp(raw.updatedAt, 'updatedAt');
-  const expiresAt =
-    raw.expiresAt === undefined ? undefined : normalizeTimestamp(raw.expiresAt, 'expiresAt');
-  const reason = normalizeReason(raw.reason);
-  return {
-    id: normalizeHoldId(raw.id),
-    scope: normalizeScope(raw.scope),
-    ...(reason ? { reason } : {}),
-    createdAt,
-    updatedAt,
-    ...(expiresAt === undefined ? {} : { expiresAt }),
-  };
-}
-
-function normalizeScope(scope: HumanControlHoldScope): HumanControlHoldScope {
-  if (!scope || typeof scope !== 'object') {
-    throw new AppError('INVALID_ARGS', 'Human-control hold requires a device scope.');
-  }
-  const deviceName = normalizeOptionalLabel(scope.deviceName, 'device name');
-  const platform = normalizeOptionalLabel(scope.platform, 'platform');
-  const kind = normalizeOptionalLabel(scope.kind, 'device kind');
-  return {
-    deviceKey: normalizeDeviceKey(scope.deviceKey),
-    ...(deviceName ? { deviceName } : {}),
-    ...(platform ? { platform } : {}),
-    ...(kind ? { kind } : {}),
-  };
-}
-
-function normalizeHoldId(id: string): string {
-  const value = typeof id === 'string' ? id.trim() : '';
-  if (!/^[a-zA-Z0-9._-]{1,128}$/.test(value)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      'Invalid human-control hold id. Use 1-128 chars: letters, numbers, dot, underscore, hyphen.',
-    );
-  }
-  return value;
-}
-
-function normalizeDeviceKey(deviceKey: string): string {
-  const value = typeof deviceKey === 'string' ? deviceKey.trim() : '';
-  if (!value || value.length > 256 || !/^[\x20-\x7E]+$/.test(value)) {
-    throw new AppError('INVALID_ARGS', 'Invalid device key. Use 1-256 printable characters.');
-  }
-  return value;
-}
-
 function normalizeDeviceAliases(deviceKeys: readonly string[]): string[] {
   return Array.from(
     new Set(
@@ -278,45 +190,4 @@ function normalizeDeviceAliases(deviceKeys: readonly string[]): string[] {
 
 function normalizeDeviceAlias(value: string): string {
   return value.trim().toLocaleLowerCase('en-US');
-}
-
-function normalizeOptionalLabel(value: string | undefined, label: string): string | undefined {
-  if (value === undefined) return undefined;
-  const normalized = value.trim();
-  if (!normalized || normalized.length > 256) {
-    throw new AppError('INVALID_ARGS', `Invalid ${label}. Use 1-256 characters.`);
-  }
-  return normalized;
-}
-
-function normalizeReason(reason: string | undefined): string | undefined {
-  if (reason === undefined) return undefined;
-  const value = reason.trim();
-  if (!value) return undefined;
-  if (value.length > 512) {
-    throw new AppError('INVALID_ARGS', 'Human-control reason must be at most 512 characters.');
-  }
-  return value;
-}
-
-function normalizeTtlMs(ttlMs: number | undefined): number | undefined {
-  if (ttlMs === undefined) return undefined;
-  if (!Number.isInteger(ttlMs) || ttlMs < MIN_HOLD_TTL_MS || ttlMs > MAX_HOLD_TTL_MS) {
-    throw new AppError(
-      'INVALID_ARGS',
-      `Human-control ttlMs must be between ${String(MIN_HOLD_TTL_MS)} and ${String(MAX_HOLD_TTL_MS)}.`,
-    );
-  }
-  return ttlMs;
-}
-
-function normalizeTimestamp(value: number, field: string): number {
-  if (!Number.isFinite(value) || value < 0) {
-    throw new AppError('COMMAND_FAILED', `Persisted human-control ${field} is invalid.`);
-  }
-  return value;
-}
-
-function cloneHold(hold: HumanControlHold): HumanControlHold {
-  return { ...hold, scope: { ...hold.scope } };
 }
